requirements.txt
Flask==3.1.2
gunicorn==22.0.0
requests==2.32.3
mysql-connector-python==9.1.0
python-dotenv==1.0.1
openai==1.59.7
